requirements.txt
python-telegram-bot==20.7
pytz==2024.1
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
